+++ src/splslda.jl
@@ -31,7 +31,7 @@ Keyword arguments:
     in the PLS computation.
 
 Same as function `plslda` (PLS-LDA) except that 
-a sparse PLSR (function `splskern`), instead of a 
+a sparse PLSR (function `splsr`), instead of a 
 PLSR (function `plskern`), is run on the Y-dummy table. 
 
 ## Examples
@@ -100,7 +100,7 @@ function splslda(X, y, weights::Weight; kwargs...)
     @assert par.nlv >= 1 "Argument 'nlv' must be in >= 1"   
     res = dummy(y)
     ni = tab(y).vals
-    embfitm = splskern(X, res.Y, weights; kwargs...)
+    embfitm = splsr(X, res.Y, weights; kwargs...)
     dafitm = list(Lda, par.nlv)
     @inbounds for i = 1:par.nlv
         dafitm[i] = lda(embfitm.T[:, 1:i], y, weights; kwargs...)